Moved some event related functions to utils
